# clients/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "node-flow",
    "version": "1",
    "description": "Workflow orchestration. Definitions are declarative JSON; workers are language-agnostic and lease tasks from a queue."
  },
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived, revocable, scoped. For CLIs and CI."
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived token from POST /v1/auth/token. For worker fleets: a captured token is useless within the hour."
      }
    }
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/v1/ns/{ns}/integrations": {
      "get": {
        "operationId": "integration_list",
        "summary": "List integrations",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "workflows:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "LLM",
                "MCP",
                "HTTP"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/integrations/{name}": {
      "delete": {
        "operationId": "integration_remove",
        "summary": "Delete an integration",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      },
      "get": {
        "operationId": "integration_get",
        "summary": "Fetch an integration",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "workflows:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      },
      "put": {
        "operationId": "integration_put",
        "summary": "Create or replace an integration",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "admin"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": [
                      "LLM",
                      "MCP",
                      "HTTP"
                    ]
                  },
                  "provider": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "description": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 1000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "baseUrl": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 2000
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "apiKeySecret": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 200
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "models": {
                    "maxItems": 100,
                    "type": "array",
                    "items": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200
                    }
                  },
                  "config": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "kind",
                  "provider"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/integrations/{name}/operations": {
      "get": {
        "operationId": "integration_operations",
        "summary": "List the operations an HTTP service publishes",
        "description": "Reads the integration’s OpenAPI document, inline or by URL.",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "workflows:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/integrations/{name}/test": {
      "post": {
        "operationId": "integration_test",
        "summary": "Check an integration works",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "workflows:write"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "model": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
    "/v1/ns/{ns}/prompts": {
      "get": {
        "operationId": "prompt_list",
        "summary": "List prompts (latest versions)",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "workflows:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
    "/v1/ns/{ns}/prompts/-/test": {
      "post": {
        "operationId": "prompt_test",
        "summary": "Run a prompt against a model",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "workflows:write"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "llmProvider": {
                    "type": "string",
                    "minLength": 1
                  },
                  "model": {
                    "type": "string",
                    "minLength": 1
                  },
                  "temperature": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 2
                  },
                  "maxTokens": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 32000
                  },
                  "template": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100000
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "version": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "variables": {
                    "type": "object",
                    "propertyNames": {
                      "type": "string"
                    },
                    "additionalProperties": {}
                  },
                  "instructions": {
                    "type": "string",
                    "maxLength": 10000
                  }
                },
                "required": [
                  "llmProvider"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
    "/v1/ns/{ns}/vector-indexes": {
      "get": {
        "operationId": "vectorIndex_list",
        "summary": "List vector indexes",
        "tags": [
          "ai"
        ],
        "x-required-scopes": [
          "workflows:read"
        ],
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Namespace slug. Must match the namespace of the credential, or the request is refused — it does not select one."
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "No or invalid credentials"
          },
          "403": {
            "description": "Authenticated, but missing a required scope"
          }
        }
      }
    },
...
  }
}